After lots of operational knowledge, change bordeaux to more red in JuliaDynamics
Unfortunately the bordeaux was too dark and difficult to discern from the black from far away. When overlaced on top of the purple it was also hard to distinguish. This new color still keeps all colorblind distinguishability but is also more distinguishable from purple. It is not very good for grey scale printing but I'll take it.